blog create done

+++ src/components/Dashboard/Blog/BlogList/index.tsx
@@ -1,7 +1,6 @@
-import React from "react";
+import React, { FormEvent } from "react";
 import "./style.css";
 import Sidebar from "../../Sidebar";
-import AddBlog from "../AddBlog";
 import {
   getFirestore,
   collection,
@@ -11,9 +10,11 @@ import {
   getDoc,
   updateDoc,
   deleteDoc,
-  doc, 
+  doc,
 } from "firebase/firestore";
 import { firebaseDatabase } from "../../../../database/firebaseConfig";
+import { ToastContainer, toast } from "react-toastify";
+import AddBlog from "../AddBlog";
 
 type BlogListDataType = {
   id: string;
@@ -24,24 +25,36 @@ type BlogListDataType = {
   date: string;
 };
 
+const initialData: BlogListDataType = {
+  id: "",
+  title: "",
+  description: "",
+  blogImage: "",
+  icon: "",
+  date: "",
+};
 const BlogList: React.FC = () => {
   const [blogItem, setBlogItem] = React.useState<BlogListDataType[]>([]);
-  const [formTitle, setFormTitle] = React.useState<string>(""); 
-  const [id, setId] = React.useState<string>(""); 
-  const [title, setTitle] = React.useState<string>(""); 
+  const [formTitle, setFormTitle] = React.useState<string>("");
+  const [ids, setIds] = React.useState<string>("");
+  const [title, setTitle] = React.useState<string>("");
   const [isLoading, setIsLoading] = React.useState<Boolean>(true);
-  const handleOpenClick = () => {
-    setFormTitle("Add Blog");
-    (document.getElementById("modal") as HTMLInputElement).style.display =
-      "block";
+  const [formReset, setFormReset] = React.useState<Boolean>(false);
+  const [modalOpen, setModalOpen] = React.useState<Boolean>(false);
+  // const [modalClose, setModalClose] = React.useState<Boolean>(false);
+  const [add, setAdd] = React.useState<Boolean>(false);
+  const [edit, setEdit] = React.useState<Boolean>(false);
+
+  const handleModalOpen = () => {
+    setModalOpen(true);
+    // setModalClose(true);
+    console.log("open: ", modalOpen);
   };
-  const handleCloseClick = () => {
-    (document.getElementById("modal") as HTMLInputElement).style.display =
-      "none";
+  const handleModalClose = () => {
+    setModalOpen(false);
+    // setModalClose(false);
+    // console.log("close: ", modalClose);
   };
-  const handleCloseClickEdit =()=>{   
-    (document.getElementById("editModal") as HTMLInputElement).style.display="none";
-  } 
 
   const getData = async () => {
     const colRef = collection(firebaseDatabase, "blog");
@@ -49,8 +62,8 @@ const BlogList: React.FC = () => {
       const result = await getDocs(colRef);
       const prepareData = result?.docs.map((item) => {
         let temp = item.data();
-        let obj: BlogListDataType  = {       
-          id: temp.id,   
+        let obj: BlogListDataType = {
+          id: temp.id,
           title: temp.title,
           description: temp.description,
           blogImage: temp.blogImage,
@@ -59,7 +72,7 @@ const BlogList: React.FC = () => {
         };
         return obj;
       });
-      setBlogItem(prepareData);   
+      setBlogItem(prepareData);
       setIsLoading(true);
       return prepareData;
     } catch (error) {
@@ -67,21 +80,28 @@ const BlogList: React.FC = () => {
     }
   };
 
-  const handleDelete = (id:string) =>{
-    const db = getFirestore(); 
-    const blogId = id.toString();
-    const docRef = doc(db, "blog", `${blogId}`);    
-
-    deleteDoc(docRef)
-    .then(() => {
-        console.log("One blog post has been deleted successfully.")
-        alert("Blog post is deleted");
-        setIsLoading(false);
-    })
-    .catch(error => {
-        console.log(error);
-    })
-  }
+  const handleDelete = (id: string) => {
+    var val = window.confirm("Are you sure to delete?");
+    if (val === true) {
+      const db = getFirestore();
+      const blogId = id.toString();
+      const docRef = doc(db, "blog", `${blogId}`);
+      deleteDoc(docRef)
+        .then(() => {
+          console.log("One Blog has been deleted successfully.");
+          setIsLoading(false);
+          const notifyDelete = () => toast("Blog is deleted");
+          notifyDelete();
+        })
+        .catch((error) => {
+          console.log(error);
+        });
+      return true;
+    } else {
+      console.log("Process Aborted");
+      return false;
+    }
+  };
 
   React.useEffect(() => {
     getData();
@@ -90,82 +110,115 @@ const BlogList: React.FC = () => {
   return (
     <React.Fragment>
       <Sidebar />
-      <section className="bloglist">
-        <div className="bloglist__row">
-          <h3 className="bloglist__row__title">Blog List</h3>
-          <div className="bloglist__row__button">
+      <section className="blogList">
+        <ToastContainer />
+        <div className="blogList__row">
+          <h3 className="blogList__row__title">Blog list</h3>
+          <div className="blogList__row__button">
             <button
-              className="bloglist__row__button__btn"
-              onClick={handleOpenClick}
+              className="blogList__row__button__btn"
+              onClick={() => {
+                handleModalOpen();
+                setAdd(true);
+              }}
             >
               + add
             </button>
-
-            <div id="modal" className="bloglist__row__modal">
-              <div className="bloglist__row__modal__content">
-                <span
-                  className="bloglist__row__modal__content__close"
-                  onClick={handleCloseClick}
-                >
-                  &times;
-                </span>
-              
-                <AddBlog formTitle={formTitle} setFormTitle={setFormTitle}
-                 setIsLoading={setIsLoading}
-                />
+            {add && modalOpen && (
+              <div id="modal" className="blogList__row__modal">
+                <div className="blogList__row__modal__content">
+                  <span
+                    className="blogList__row__modal__content__close"
+                    onClick={() => {
+                      setFormReset(true);
+                      handleModalClose();
+                      setAdd(false);
+                    }}
+                  >
+                    &times;
+                  </span>
+                  <AddBlog
+                    formTitle="Add Blog"
+                    setFormTitle={setFormTitle}
+                    setIsLoading={setIsLoading}
+                    formReset={formReset}
+                    setFormReset={setFormReset}
+                    setModalOpen={setModalOpen}
+                  />
+                </div>
               </div>
-            </div>
+            )}
           </div>
-          <table className="bloglist__row__table">
-            <tr className="bloglist__row__table__row">
-              <th className="bloglist__row__table__row__text">Title</th>
-              <th className="bloglist__row__table__row__text">
-                Description
-              </th>
-              <th className="bloglist__row__table__row__text">Date</th>
-              <th className="bloglist__row__table__row__text">Actions</th>
+          <table className="blogList__row__table">
+            <tr className="blogList__row__table__row">
+              <th className="blogList__row__table__row__text">Title</th>
+              <th className="blogList__row__table__row__text">Image</th>
+              <th className="blogList__row__table__row__text">Description</th>
+              <th className="blogList__row__table__row__text">Actions</th>
             </tr>
-
             {blogItem?.map((blog) => {
               return (
-                <tr className="bloglist__row__table__row">
-                  <td className="bloglist__row__table__row__text">
+                <tr className="blogList__row__table__row" key={blog?.id}>
+                  <td className="blogList__row__table__row__text">
                     {blog.title}
                   </td>
-                  <td className="bloglist__row__table__row__text">
-                    {blog.description}
+                  <td className="blogList__row__table__row__text">
+                    <img
+                      height="50px"
+                      width="50px"
+                      src={blog.blogImage}
+                      alt="Blog Images"
+                    />
                   </td>
-                  <td className="bloglist__row__table__row__text">
-                    {blog.date}
+                  <td className="blogList__row__table__row__text">
+                    {blog.description.slice(0, 85)}
                   </td>
-                  <td className="bloglist__row__table__row__text">
-                    <button className="bloglist__row__table__row__button__edit"
-                     onClick={
-                      ()=>{
-                        setFormTitle("Edit Blog");                                         
-                        setId(blog.id);
+                  <td className="blogList__row__table__row__text">
+                    <button
+                      className="blogList__row__table__row__button__edit"
+                      onClick={() => {
+                        setFormTitle("Edit Blog");
+                        setIds(blog.id);
                         setTitle(blog.title);
-                        (document.getElementById("editModal") as HTMLInputElement).style.display="block";  
-                      }
-                     } 
+                        setEdit(true);
+                        setAdd(false);
+                        setModalOpen(true);
+                      }}
                     >
                       edit
                     </button>
-                    <div id="editModal" className="bloglist__row__modal"> 
-                                      <div className="bloglist__row__modal__content">
-                                          <span className="bloglist__row__modal__content__close" 
-                                          onClick={handleCloseClickEdit}
-                                          >&times;</span>                
-                                          <AddBlog formTitle={formTitle} 
-                                            setFormTitle={setFormTitle} 
-                                            ids={id} 
-                                            titleForm={title}
-                                            setIsLoading={setIsLoading}
-                                             />
-                                      </div>
-                                  </div>
-                    <button className="bloglist__row__table__row__button__delete"
-                     onClick = {()=> handleDelete(blog.id)}
+                    {edit && modalOpen && (
+                      <div
+                        id="editModal"
+                        className="blogList__row__modal"
+                        style={{ backgroundColor: "rgba(0, 0, 0, 0.08)" }}
+                      >
+                        <div className="blogList__row__modal__content">
+                          <span
+                            className="blogList__row__modal__content__close"
+                            onClick={() => {
+                              handleModalClose();
+                              setEdit(false);
+                            }}
+                          >
+                            &times;
+                          </span>
+                          <AddBlog
+                            formTitle="Edit Blog"
+                            setFormTitle={setFormTitle}
+                            ids={ids}
+                            titleForm={title}
+                            setIsLoading={setIsLoading}
+                            formReset={formReset}
+                            setFormReset={setFormReset}
+                            setModalOpen={setModalOpen}
+                          />
+                        </div>
+                      </div>
+                    )}
+                    <button
+                      className="blogList__row__table__row__button__delete"
+                      onClick={() => handleDelete(blog.id)}
                     >
                       delete
                     </button>
